Revert "Merge pull request #2 from hokitlee/master"
This reverts commit 2285e80d67c85b167d8dd39c0d9d7feafbb5f63d, reversing
changes made to f5c0e559c04d8c2f64002c0acfdccf93ceb33ebb.
