+++ commerce_web/src/views/Home.vue
@@ -1,5 +1,5 @@
 <template>
-  <div id="home" class="container-fluid">
+  <div class="home">
       <Heander></Heander>
       <div class="row">
           <div class="col-md-2">
